階段 4 開工：覆蓋矩陣、Tauri 殼、Rust→TS 型別產生
## 覆蓋矩陣（loom-core/src/coverage.rs）

wireframe 對過的主畫面：每條邏輯連線 × 每個環境，空格就是漏洞。

它刻意**不重新定義什麼叫缺漏**——狀態一律從 lint 的結果推導。
自己判斷一次的話，遲早會跟 lint 面板講出不一樣的話，
那使用者就不會再相信任何一邊。這裡唯一自己算的是數字（幾段、幾台）。

九個測試，其中最重要的一條是「矩陣的狀態不會跟 lint 面板互相矛盾」：
刻意弄壞三個環境的不同地方，逐格比對兩邊的說法。

## EnvIndex 抽成共用模組

原本藏在 lint.rs 裡。覆蓋矩陣也要做萬用字元比對，不共用的話
會把剛修掉的平方級問題再引入一次。

## Tauri 殼（src-tauri/）

薄。三個 command：open_project / recheck / save_project，
每個都短到一眼看得出它沒有偷做決定。

一次回傳 Snapshot（模型 + findings + 矩陣）而不是分三次問——
分開問會有中間狀態，畫面會短暫地自相矛盾。

## 型別產生（tauri-specta）

src/lib/bindings.ts 由 Rust 自動產生，連文件註解都帶過去。
產生器與 App 共用同一個 builder()，不可能出現「command 加了但型別沒更新」。

兩個坑：
- specta 2 / tauri-specta 2 只有 rc 版（2.0.0-rc.25），靠 Cargo.lock 鎖住
- specta 拒絕匯出 usize（怕 BigInt 精度），0.0.12 沒有全域開關。
  所以跨邊界的計數改用 u32——「期望有幾台機器」本來也不需要 64 位元

## 紀律

src-tauri 進了同一個 workspace 之後，把 tauri 加進 loom-core 會「剛好能編」，
不會有任何錯誤提醒。新增 `mise run guard` 守這條線，已驗證它真的會擋下來。

test:core 也改成只跑 loom-core（1.5 秒）——加上 --workspace 會連 Tauri
一起編，那一層就不再是秒級的了。

+++ crates/loom-core/src/environment.rs
@@ -10,6 +10,7 @@ use crate::logical::Protocol;
 
 /// 運算載體的種類。
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 #[serde(rename_all = "kebab-case")]
 pub enum NodeKind {
     Physical,
@@ -19,6 +20,7 @@ pub enum NodeKind {
 
 /// Endpoint 在某環境的實際樣貌：定義加上位址。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct Endpoint {
     pub id: Id,
     pub slug: String,
@@ -37,6 +39,7 @@ pub struct Endpoint {
 /// 叢集就是多個 Instance：12 台 VM 各跑一個 Redis process
 /// 就是 12 個 `ContainerInstance`。節點數不另外存數字，避免兩份資料不一致。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct ContainerInstance {
     pub id: Id,
     pub slug: String,
@@ -51,6 +54,7 @@ pub struct ContainerInstance {
 
 /// 機器：實體機、VM、Linux container。C4 的 `Deployment Node`，可巢狀。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct DeploymentNode {
     pub id: Id,
     pub slug: String,
@@ -78,6 +82,7 @@ impl DeploymentNode {
 /// 例如金流系統：prod 用正式閘道，test 用 sandbox。
 /// 它**不放在 [`DeploymentNode`] 底下**——那些機器不是我們的，我們只知道位址。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct SoftwareSystemInstance {
     pub id: Id,
     pub slug: String,
@@ -94,6 +99,7 @@ pub struct SoftwareSystemInstance {
 /// 它**不含 Container**，但有自己的 endpoint（VIP 位址），
 /// 因此連線可以「經過」它——這也是為什麼經過 F5 的流量會拆成兩段。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct InfrastructureNode {
     pub id: Id,
     pub slug: String,
@@ -103,6 +109,7 @@ pub struct InfrastructureNode {
 
 /// 連線一端指向的 Instance：可以是一個，也可以是一整群。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 #[serde(rename_all = "kebab-case")]
 pub enum InstanceRef {
     /// 指名一個 Instance。
@@ -115,7 +122,9 @@ pub enum InstanceRef {
     Pattern {
         slug_pattern: String,
         #[serde(default, skip_serializing_if = "Option::is_none")]
-        expect: Option<usize>,
+        /// 用 `u32` 而不是 `usize`：這是「期望有幾台機器」，不是記憶體索引。
+        /// 而且 `usize` 不能匯出成 TypeScript（specta 怕 BigInt 精度問題）。
+        expect: Option<u32>,
     },
 }
 
@@ -133,6 +142,7 @@ pub enum InstanceRef {
 ///     endpoint: ep-f5-redis
 /// ```
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 #[serde(rename_all = "kebab-case")]
 pub enum Endpointing {
     Instance {
@@ -174,6 +184,7 @@ pub enum Endpointing {
 /// `serves` 指向它所服務的邏輯 [`Relationship`](crate::logical::Relationship)。
 /// Lint 把貼同一個 `serves` 的連線攤開成一張圖，檢查從來源走不走得到目標。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct Connection {
     pub id: Id,
     pub serves: Id,
@@ -185,6 +196,7 @@ pub struct Connection {
 
 /// 一個部署環境。
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct Environment {
     pub id: Id,
     pub slug: String,
