+++ crates/loom-core/tests/lint_scale.rs
@@ -151,7 +151,7 @@ fn environment(scale: Scale, e: usize) -> Environment {
                 from: Endpointing::Instance {
                     target: InstanceRef::Pattern {
                         slug_pattern: format!("svc{i:03}-*"),
-                        expect: Some(scale.nodes),
+                        expect: Some(scale.nodes as u32),
                     },
                     endpoint: None,
                 },
@@ -171,7 +171,7 @@ fn environment(scale: Scale, e: usize) -> Environment {
                 to: Endpointing::Instance {
                     target: InstanceRef::Pattern {
                         slug_pattern: format!("svc{next:03}-*"),
-                        expect: Some(scale.nodes),
+                        expect: Some(scale.nodes as u32),
                     },
                     endpoint: Some(Id::new(def(next))),
                 },
